+++ NEWS.md
@@ -1,7 +1,8 @@
-# palaeoverse 1.1.1.900
+# palaeoverse (development version)
 
 * Fixed unnecessary column output in palaeorotate (#78)
 * Fixed prior updates to "point" method in palaeorotate which introduced binding issues for some specific datasets
+* Added additional tests for palaeorotate
 
 # palaeoverse 1.1.1